Added a csv reader
Create a function that reads csv files based on the one used in the CBP computer project. The list of names is assigned to the values. If any addition parameters are added or changed it has to be done in the csv file and the code. Also fix some PEP issues.