Changed component names
Changed the component names and added sulfur. I just set the stoichiometric coefficient of sulfur to zero in the matrix and also adjusted the relevant initial conditions so that there is no sulfur present at the moment. I haven’t read about the actual reactions taking place and will need to review that to specify the stoichiometric coefficients for all the components.