Added variables to csv and added component names
Combined the variables of the Gustafsson model and the Andersson model. The top two variables in the spreadsheet are Andersson_model and Gustafsson_model. These need to be set to either 0 or 1, depending on which model you want to use. These new variable names should be kept in mind when writing new code and if a new variable is desired it should be stated in the parameters.csv file.