Data Handling Overhaul
I have modified the reader function to now create a “dirc” dictionary where all the names are assigned to the respective values. I also replaced almost every parameter with the new parameters dictionary. I haven’t replaced the D and K values because the diffusivity function has not been merged. The names and values can still be changed, it’s just the way the data is handled that is changed.