Added timer
Added code to determine the time it takes the program to run. I think this will help to test various methods and the effects of for loops etc. I put it before the plotting code because when put after plt.show() line the program only resumes after the graph window has been closed. I also assumed that the plotting will always take more or less the same time.