Reconfigured initial conditions
The initial conditions is reconfigured to better match the pulping process. First the reactions was modified to include the alkali in the reaction rate expression. The stoichiometric table was changed so that each reaction rate only decreases one component. The alkali does not yet decrease with time. The lignin and carbohydrates can no longer diffuse inside the wood chip or from the wood chip. The initial lignin and carbohydrate content in the wood chip is also set.