Model selection method test
I modified the reaction_rates function. It now has two terms for each reaction rate constant. If one model is chosen the corresponding coefficient is set to 1 and the other model’s coefficient is set to 0. Don’t know if this model is efficient but I cannot think of any other method to handle different models at the moment. Also added sulfur to the initial liquor.