Added exponential function
Added a function to calculate the Gustafsson exponential constants in the reaction rates. This can be used for normal Arrhenius type terms but the first input will have to be zero and the second a negative number. I think it would be better to have a separate function for normal exponential terms.