Added Diffusivity function
I added a diffusivity function (fick_constant) to determine the Fick’s diffusion constants. The function does not calculate anything yet but only gives values back. I think we will need to do something like this for the Fick’s – and external diffusion constants. The only components we care about in the system that will diffuse is the alkali and sulfur. If they are always last on the list of components this function will work for both models as all other components will not be able to diffuse. The same method for handling the different models as in the reaction_rates can be used here. Also fixed the mass_frac index.